Address review: unbounded printf, shutdown ordering, unsigned delta
Three things the review caught.

The documented printf workaround is itself unbounded: %.*s takes an int
precision, so a result over INT_MAX passes a negative one, which printf reads as
"no precision" and goes looking for a NUL. chdb.h now documents fwrite, and
chdbStub uses it.

Context::shutdown() released the global holders before shared->shutdown(), but
ContextSharedPart::shutdown() reaches code that reads
Context::getGlobalContextInstance() -- StreamingStorageRegistry's shutdown hands
it to DatabaseCatalog::tryGetTable() -- so a server shutting down with a
registered queue table would have been handed a null context. Released on the way
out instead, via a scope guard rather than a statement after the call: an
exception on the way through shutdown must still leave the holders released, or
the next engine start in the process hits the assertion again.

The C reproducer computed its per-cycle delta as (double)(after - before), which
wraps into a huge positive number if the heap happened to shrink. Converted
before subtracting.

+++ examples/chdbAsanRegressionTest.c
@@ -151,7 +151,9 @@ static void test_engine_start_leak_budget(void)
     }
     size_t after = live_bytes();
 
-    double per_cycle = (double)(after - before) / (double)cycles;
+    /* Convert before subtracting: an unrelated free between the two reads would make
+     * a size_t subtraction wrap into a huge positive delta. */
+    double per_cycle = ((double)after - (double)before) / (double)cycles;
     printf("  %.0f bytes retained per connect/close cycle (budget %d)\n", per_cycle, LEAK_BUDGET_PER_CYCLE);
     CHECK(per_cycle < (double)LEAK_BUDGET_PER_CYCLE, "engine start/stop stays within its leak budget");
 }
